# src/api-spec/ontology.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tellus Ontology Engine \u2014 Full API",
    "description": "Comprehensive API specification for the Tellus Ontology Engine. Covers ontology CRUD, object types, properties, link types, and action types.",
    "version": "2.0.0",
    "contact": {
      "name": "Tellus Engineering",
      "email": "engineering@tellus.dev"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v2/ontologies/{ontologyId}/objectTypes": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OntologyId"
        }
      ],
      "post": {
        "operationId": "createObjectType",
        "summary": "Create a new object type",
        "description": "Creates a new object type within the specified ontology.",
        "tags": [
          "Object Types"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "apiName",
                  "displayName"
                ],
                "properties": {
                  "apiName": {
                    "type": "string",
                    "description": "Unique programmatic name (e.g. 'customer', 'order')",
                    "example": "customer"
                  },
                  "displayName": {
                    "type": "string",
                    "description": "Human-readable name",
                    "example": "Customer"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional description"
                  },
                  "icon": {
                    "type": "string",
                    "nullable": true,
                    "description": "Icon identifier (e.g. an emoji or icon library key)"
                  },
                  "iconColor": {
                    "type": "string",
                    "nullable": true,
                    "description": "Icon color (hex or named color)"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "deprecated",
                      "experimental"
                    ],
                    "default": "active",
                    "description": "Lifecycle status of the object type"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Object type created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectType"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Object type with that apiName already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listObjectTypes",
        "summary": "List object types",
        "description": "Returns all object types within the specified ontology.",
        "tags": [
          "Object Types"
        ],
        "responses": {
          "200": {
            "description": "A list of object types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ObjectType"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/objectTypes/{apiName}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OntologyId"
        },
        {
          "$ref": "#/components/parameters/ObjectTypeApiName"
        }
      ],
      "get": {
        "operationId": "getObjectType",
        "summary": "Get a single object type",
        "description": "Returns the object type identified by apiName within the given ontology.",
        "tags": [
          "Object Types"
        ],
        "responses": {
          "200": {
            "description": "The requested object type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectType"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateObjectType",
        "summary": "Update an object type",
        "description": "Updates an existing object type's display name, description, icon, or status.",
        "tags": [
          "Object Types"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "icon": {
                    "type": "string",
                    "nullable": true
                  },
                  "iconColor": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "deprecated",
                      "experimental"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Object type updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectType"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteObjectType",
        "summary": "Delete an object type",
        "description": "Permanently deletes the object type and all its properties.",
        "tags": [
          "Object Types"
        ],
        "responses": {
          "204": {
            "description": "Object type deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v2/ontologies/{ontologyId}/linkTypes": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OntologyId"
        }
      ],
      "post": {
        "operationId": "createLinkType",
        "summary": "Create a new link type",
        "description": "Creates a new link type connecting two object types within the specified ontology.",
        "tags": [
          "Link Types"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "apiName",
                  "displayName",
                  "cardinality",
                  "sourceObjectTypeApiName",
                  "targetObjectTypeApiName"
                ],
                "properties": {
                  "apiName": {
                    "type": "string",
                    "description": "Unique programmatic name for the link type",
                    "example": "customer_orders"
                  },
                  "displayName": {
                    "type": "string",
                    "description": "Human-readable name",
                    "example": "Customer Orders"
                  },
                  "cardinality": {
                    "type": "string",
                    "enum": [
                      "ONE_TO_ONE",
                      "ONE_TO_MANY",
                      "MANY_TO_ONE",
                      "MANY_TO_MANY"
                    ],
                    "description": "Relationship cardinality between source and target object types"
                  },
                  "sourceObjectTypeApiName": {
                    "type": "string",
                    "description": "apiName of the source object type",
                    "example": "customer"
                  },
                  "targetObjectTypeApiName": {
                    "type": "string",
                    "description": "apiName of the target object type",
                    "example": "order"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional description of the link type"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link type created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkType"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Link type with that apiName already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listLinkTypes",
        "summary": "List link types",
        "description": "Returns all link types within the specified ontology.",
        "tags": [
          "Link Types"
        ],
        "responses": {
          "200": {
            "description": "A list of link types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LinkType"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v2/ontologies/{ontologyId}/objectTypes/{apiName}/datasource": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OntologyId"
        },
        {
          "$ref": "#/components/parameters/ObjectTypeApiName"
        }
      ],
      "get": {
        "operationId": "getDatasource",
        "summary": "Get datasource configuration",
        "description": "Returns the registered datasource for the specified object type.",
        "tags": [
          "Datasources"
        ],
        "responses": {
          "200": {
            "description": "Datasource configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "operationId": "registerDatasource",
        "summary": "Register a datasource",
        "description": "Registers or updates the backing datasource for the specified object type.",
        "tags": [
          "Datasources"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "datasetRid": {
                    "type": "string",
                    "description": "Dataset resource identifier"
                  },
                  "columnMapping": {
                    "type": "object",
                    "description": "Mapping of object type properties to dataset columns"
                  },
                  "primaryKeyColumn": {
                    "type": "string",
                    "description": "Column used as primary key"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Datasource registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "OntologyId": {
        "name": "ontologyId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "UUID of the ontology"
      },
...
    },
    "schemas": {
...
      "ObjectType": {
        "type": "object",
        "properties": {
          "objectTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "ontologyId": {
            "type": "string",
            "format": "uuid"
          },
          "apiName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "iconColor": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "deprecated",
              "experimental"
            ]
          },
          "propertyCount": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}